fixed compilation errors

+++ workflows/src/main/kotlin/com/r3/developers/csdetemplate/digitalcurrency/workflows/SellMortgageFlow.kt
@@ -108,7 +108,7 @@ class FinalizeSellMortgageResponderFlow: AbstractFlow(), ResponderFlow {
 
             val availableTokens = ledgerService.findUnconsumedStatesByType(DigitalCurrency::class.java)
             val coinSelection = CoinSelection()
-            val (currencyToSpend, spentCurrency) = coinSelection.selectTokensForTransfer(price.toInt(),
+            val (currencyToSpend, spentCurrency) = coinSelection.selectTokensForTransfer(price.toDouble(),
                                                             buyer.ledgerKeys.first(),
                                                             seller.ledgerKeys.first(),
                                                             availableTokens)
